Refactor : Comment 도메인 관련 파일 수정 1. 간단한 API 테스트를 위한 Controller와 Service에 method 생성 2. Repository에서 파일 위치 변경에 따른 import 수정 3. Field id 컬럼명 수정
